+++ Assets/ViveGrip/Examples/Scripts/ViveGripExample_Manager.cs
@@ -1,12 +1,22 @@
 using UnityEngine;
 using System.Collections.Generic;
+using ViveGrip.TypeReferences;
 
 // See EXTENSIONS.md for more information
 
 public class ViveGripExample_Manager : MonoBehaviour {
   public bool disableAllHighlighting = false;
+  [ClassImplements(typeof(ViveGrip_HighlightEffect))]
+  public ClassTypeReference highlightEffectForEverything = null;
+  public bool changeHighlightsAtStart = false;
   public float gripStrengthMultiplier = 1f;
   public float gripRotationMultiplier = 1f;
+
+  void Start() {
+    if (changeHighlightsAtStart) {
+      ChangeAllHighlightEffects();
+    }
+  }
 
   void Update() {
     SetHighlighting();
@@ -14,7 +24,7 @@ public class ViveGripExample_Manager : MonoBehaviour {
   }
 
   void SetHighlighting() {
-    foreach(ViveGrip_Highlighter highlighter in FindObjectsOfType<ViveGrip_Highlighter>()) {
+    foreach(ViveGrip_Highlighter highlighter in AllHighlighters()) {
       highlighter.enabled = !disableAllHighlighting;
     }
   }
@@ -22,5 +32,21 @@ public class ViveGripExample_Manager : MonoBehaviour {
   void SetGripStrength() {
     ViveGrip_JointFactory.LINEAR_DRIVE_MULTIPLIER = gripStrengthMultiplier;
     ViveGrip_JointFactory.ANGULAR_DRIVE_MULTIPLIER = gripRotationMultiplier;
+  }
+
+  void OnValidate() {
+    ChangeAllHighlightEffects();
+  }
+
+  void ChangeAllHighlightEffects() {
+    if (AllHighlighters().Length == 0) { return; }
+    Debug.Log("Changing all highlight effects to: " + highlightEffectForEverything.ToString());
+    foreach(ViveGrip_Highlighter highlighter in AllHighlighters()) {
+      highlighter.UpdateEffect(highlightEffectForEverything.Type);
+    }
+  }
+
+  ViveGrip_Highlighter[] AllHighlighters() {
+    return FindObjectsOfType<ViveGrip_Highlighter>();
   }
 }